rename api_resources to endpoints
